组织架构以及滚动更新的功能

+++ backend/src/main/java/com/kakarote/ai_crm/entity/PO/ManagerUser.java
@@ -83,7 +83,7 @@ public class ManagerUser implements Serializable {
 	 * 部门
 	 */
 	@Schema(description = "部门")
-	private Integer deptId;
+	private Long deptId;
 	/**
 	 * 岗位
 	 */
